pyproject.toml
[project]
name = "customer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=24.3.0",
    "fastapi>=0.101.0",
    "httpx>=0.27.0",
    "langchain>=0.2.10",
    "langchain-core>=0.2.0",
    "langchain-groq>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "mypy>=1.5.0",
    "pydantic>=2.3.0",
    "pytest>=8.2.0",
    "pytest-cov>=4.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "ruff>=0.2.0",
    "streamlit>=1.27.0",
    "types-requests>=2.31.0",
    "typing-extensions>=4.7.1",
    "uvicorn[standard]>=0.25.0",
]
